requirement.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloud-sql-python-connector==1.18.3
cryptography==45.0.6
dnspython==2.7.0
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
frozenlist==1.7.0
google-auth==2.40.3
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.6.3
packaging==25.0
pluggy==1.6.0
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.2
PyMySQL==1.1.1
pytest==8.4.1
requests==2.32.4
rsa==4.2
SQLAlchemy==2.0.42
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.20.1
